Added file upload validation and tests
The Logo processing code now validates the uploaded file and decides if
it can be converted into an hmm logo or not. The tests check that the
code is doing the correct thing for a given set of uploaded data. This
can be modified as we find other edge cases.