Added option to redirect from POST if from browser
We want to be able to redirect the browser away from a POST if the
result is a success, so that they don't resubmit the form when refreshing
the page. This was not possible, without messing up the REST side of
things unless there is a different response for browsers.