A lot of letters were rendered that shouldn't have
All the letters that had a value between 0 and 0.01 were being rendered
after the change to allowing negative values. These were not being
scaled correctly and were being dumped into the top left corner of the
image. I changed the else statement to an else if so that we can exclude
them from the final image.