Removed posscore option and updated js.
The javascript for the logo has been updated to the latest version. This
includes changes to allow column information to be displayed in a
companion div somewhere else on the page when the column is clicked.

The posscore option was removed from the logo form as it will no longer
be drawn. Instead, score will only draw the positive values and then
display the negative values if a column is clicked.