Download scaled png only when emission heights used.
Previously the default was to download a png that had been scaled to the
max theoretical height. Since that is not useful when using score or
bild height calculations, I reversed the default to use max observed
height. I also added an option to download a scaled image if using the
emission calculations.