Added a media query to improve mobile display
When the screen size gets small, the display changed to a single column
layout that ran all the way up to the edge. This didn't look as nice as
the larger display, so I added a media query which places a border
around the content. It does remove some of the usable space for the
page, but the page looks less "crammed" into the display port.