Need to remove existing minified js.
Since we are using git to do the checkouts for this project, we dont
blow away and checkout a whole new directory like we did for subversion.
This is faster, but does leave behind build artifacts. If we dont remove
those, then they can get rolled into the next build and cause
duplications of javascript in the minified files.