Fixes a validation error reported by Tom
Tom found that an aligned fasta file was causing the site to crash,
rather than to report that there was an illegal character in the
sequence. With help from Travis, I found the root cause and updated the
validation code to report the correct error and location within the file.