Fixed AFA validation problem reported by Travis
The Easel::Validation code that was used for HMMER would punt on AFA
files as it couldn't determine if people had uploaded a lot of fasta
entries in one file to be searched individually or as a multiple sequence
alignment. Since the logo server does not run with a single fasta, this
is an easy choice to make and the code has been modified to assume the
file is aligned fasta if all the sequences are of the same length.