Added nicer error message for alignment processed hmms
When people upload an HMM the website will set the alignment processing
option to convert to an HMM if it can. When it can't things will go
wrong as the default option is to use the observed counts processing.
Since we can detect the issue on the server side, we can change the
chosen value accordingly and give the results the user would be
expecting with a message explaining what happened.

There needed to be a different error message displayed on the site than
was being displayed via the API as the way to correct the issue was
different. This was done by checking to see if we think the request is
coming from a browser or some other software.