Trying to stop overlapping letters problem.
This change reverse the order in which the letters are rendered, so the
smaller bottom letters are drawn on top of the larger letters. Since
they will cover a smaller amount of the larger letter, the overlap is
less of an issue when trying to identify the letter being covered up.