Adds POST patch to Action::REST::ForBrowsers
The Catalyst::Action::REST::ForBrowsers module only overrides the GET
method. Since we also want to override the POST method and return a
different response when received from a browser, we need to have this
module which adds the option to return content from a *_POST_html
method.

I sent an email to the maintainers of the module on CPAN, but heard
nothing back, so this will live here for now.