Added help message when css fonts are disabled.
The logos rendered with canvas look awful when rendered in a serif font.
We set the stylesheet to make sure that things are rendered using a
sans-serif font, but this can be overridden by the browser. Since there
is no way for us to stop that, I have added a javascript check that
looks to see if the font we supplied is being used. If it isn't then we
throw a warning onto the page explaining why the logo looks so
unreadable.