Add help links to settings options and remove pop up colors
This change adds help links to all the settings options, so that people
can find out more about each of them. It also removes the colors from
the column info pop up if the consensus colors are in use as the other
colors don't make any sense in that context.