Changes default processing with API to match form
This changes the default post processing value, when using the API, to
the same value as when using the web form.

Also added some additional documentation to help clear up a few rough
edges and make things clearer to the user. Primarily filled in the
blanks about the processing and letter height options.

Finally, I added some extra options to the curl examples, so that people
that aren't familiar with curl can see how to pass multiple arguments to
the server.

fixes #4
fixes #5